VirtHCK: Made Windows10 boot with "qemu64" CPU
Windows 10 tech preview (10074) did not boot with the qemu64 CPU, but rather gave BSOD with "SYSTEM_THREAD_EXCEPTION_NOT_HANDLED" error. That was true for the installation media as well.
Now, with the added CPU flag "fsgsbase" it boots correctly. The CPU options added previously (+x2apic, model=13) do not seem to make any difference, but are kept meanwhile, because they do not seem to cause any troubles as well.

Signed-off-by: Leonid Bloch <leonid@daynix.com>