VirtHCK: Added a script to automate filters update
This works in both HCK and HLK. The needed input is the download source (which
as of now is the same for both HCK and HLK) and the version of the installed kit.
The latest version of HCK as of now is 8.1, and of HLK - 10. You can check your
version by looking at the content of "C:\Program Files (x86)\Windows Kits". Notice,
that several versions can be installed simultaneously, which poses no problem.

The "Update_filters.bat" script can be kept on the desktop, for convenience, and
run once in few days, to guarantee the presence of the latest filters.

Signed-off-by: Leonid Bloch <leonid@daynix.com>