VirtHCK: Enable world access to new clients
Previously, few lines needed to be manually uncommented to grant the clients Internet access for activation purposes. After activation they needed to be commented again, manually.
Now, if at least one of the clients' images is empty (under 2MB in size) Internet access is granted, to enable activation during installation.
Also added an option to deny Internet access to clients always, even if one of them is new.
There is no option to enable access forcefully, for security reasons. If that is REALLY needed, it can be done manually.

Signed-off-by: Leonid Bloch <leonid@daynix.com>