VirtHCK: KitAutosetup: Controller and Client installation scripts changed
Now the Controller and the Client installation scripts (that are run
automatically by SYS_SETUP) are checking for availability of the
installation files, in order to reduce the chances of a silent failure.
Also, some rewording and efficiency changes are made (do not reboot if
the installation is not successful, etc.).

Signed-off-by: Leonid Bloch <leonid@daynix.com>