VirtHCK: Make the world access for the clients a well-defined option
World access for clients should be enabled ONLY for their activation.
Previously, unless specifically disabled, it was enabled if the image
was empty, assuming that starting VirtHCK with an empty image is only
needed for installation. But an empty image might also be a snapshot
of a ready-made image, or we might want to enable world access for the
clients after installation, if we forgot to activate them previously.
To make this better controlled and more straightforward, this option is
now set explicitly, in an "Options for setup installation" section,
along with "UNSAFE_CACHE".

Signed-off-by: Leonid Bloch <leonid@daynix.com>