VirtHCK: Added a possibility to specify a config file other than the default
Also some flexibility improvements that are related to the config file
specification were made. In particular, the bridge-up scripts are now
not part of the code, but are generated on start. Also, the flag file
for VM respawning was given a unique ID. These changes should allow
running several VirtHCK sessions, with different configurations, from
the same directory.

Signed-off-by: Leonid Bloch <leonid@daynix.com>