Fixing a bug with multiqueue networking devices
The [ notation does not have the =~ (regex) comparison operator, and so
queue_len_tx() returned a wrong result if several queues were needed
(always as 1 queue).

The [[ notation would have worked, but would not have been compatible
with older shells, as VirtHCK strives to be.

So a case notation is used now.

Signed-off-by: Leonid Bloch <leonid@daynix.com>