Make an option to run QEMU as a non-root user when started with sudo
VirtHCK needs to be started by root or via sudo, at the very least in
order to create the tap network interfaces. But it is a bad practice to
run QEMU as root. This patch adds a config option "RUN_QEMU_AS_ROOT",
that if set to "false", QEMU will drop the root privileges after
starting, if run via sudo. Thus all the interfaces will be setup as
root initially, but the process itself will be run by an unprivileged
user.

Signed-off-by: Leonid Bloch <leonid.bloch@esrf.fr>