Make an option to use a sufficient L2 cache for QCOW2 images
Setting a sufficient amount of L2 cache when using QCOW2 images can
improve the performance significantly. The default cache size is enough
to cover 8 GB of an image with the default cluster size of 64 KB. Since
practically all the images used with VirtHCK are larger than that, this
can be a bottleneck for performance. Hear an option is added to cover
the entire virtual size of QCOW2 images. The memory overhead is not
big: for example, to cover a 100 GB image with 64 KB cluster size,
a mere 12.5 MB of RAM will be used for L2 cache.

More details on that here:
http://git.qemu.org/?p=qemu.git;a=blob;f=docs/qcow2-cache.txt

Signed-off-by: Leonid Bloch <leonid.bloch@esrf.fr>